requirements.txt
huggingface_hub
pdf2image
unstructured-inference
pi-heif
fastapi
uvicorn
pdfplumber
unstructured
sentence-transformers
faiss-cpu
langfuse
Pillow
torch
transformers
qdrant-client
chromadb
clip
python-dotenv
streamlit
requests
